requirements/dev.txt
# -r base.txt
# flask-debugtoolbar==0.16.0
# pytest==8.4.1
# pytest-flask==1.3.0
# pytest-cov==6.0.0
# black==24.10.0
# flake8==7.1.1
# mypy==1.13.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-CORS==6.0.1
Flask-Migrate==4.1.0
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
psycopg2-binary==2.9.10
python-dotenv==1.1.1
alembic==1.16.4
flask-socketio==5.5.1
celery==5.5.3
redis==6.2.0
email-validator==2.2.0
bcrypt==4.3.0
python-dateutil==2.9.0.post0
click==8.1.7
Flask-Caching==2.1.0
redis==5.0.1